cli: move to internal package
