+++ internal/cli/common_test.go
