+++ internal/cli/image.go
