+++ internal/cli/image_list.go
