+++ internal/cli/profile_check.go
