+++ torcx.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/Sirupsen/logrus"
 
-	"github.com/coreos/torcx/cli"
+	"github.com/coreos/torcx/internal/cli"
 	"github.com/coreos/torcx/pkg/multicall"
 )
